+++ scaler/io/cpp/src/intra_process_connector.cpp
@@ -1,4 +1,4 @@
-#include "intra_process.hpp"
+#include "intra_process_connector.hpp"
 
 void IntraProcessConnector::ensure_epoll()
 {
