+++ scaler/io/cpp/src/intra_process_connector.hpp
